+++ src/flows/sign/cli.ts
@@ -1,16 +1,17 @@
+import { EntropyLogger } from "src/common/logger"
 import { initializeEntropy } from "../../common/initializeEntropy"
-import { debug } from "../../common/utils"
 import * as config from '../../config'
 import { signWithAdapters } from './sign'
 
 // TODO: revisit this file, rename as signEthTransaction?
 export async function cliSign ({ address, message, endpoint }) {
-
+  const logger = new EntropyLogger('CLI::SIGN', endpoint)
   const storedConfig = await config.get()
   const account = storedConfig.accounts.find(account => account.address === address)
   if (!account) throw Error(`No account with address ${address}`)
   // QUESTION: is throwing the right response?
-  debug('account', account)
+  logger.debug('account:')
+  logger.debug(account)
 
   const entropy = await initializeEntropy({ keyMaterial: account.data, endpoint })